+++ source/mod_parameters.F90
@@ -16,9 +16,9 @@ module mod_parameters
   integer       :: dim
   !! Dimension: 4 spins x number of atoms in the unit cell x number of orbitals^2
 
-  integer, dimension(3) :: kp_in
+  integer*8, dimension(3) :: kp_in
   !! Number of k-points in each direction
-  integer :: kptotal_in
+  integer*8 :: kptotal_in
   !! Total number of k-points
 
   !========================================================================================!
